# openapi.yaml
openapi: 3.0.1
info:
  title: SurvHey BackEnd
  description: 'This is the OAS for the BackEnd of the SurvHey Project.'
  version: 1.0.0
servers:
  - url: http://localhost:8080/v2
...
paths:
...
  /survey:
    post:
      tags:
        - survey
      summary: Create new survey
      description: Creates a new survey for the logged-in user
      operationId: createSurvey
      requestBody:
        description: Created survey object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveyPrepare'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
...
  /survey/{id}:
    get:
      tags:
        - survey
      summary: Get survey by id
      description: Returns survey by id without need for authentication. This endpoint returns the survey for end-users
      operationId: getSurveyById
      parameters:
        - name: id
          in: path
          description: ID of survey to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
        - survey
      summary: Update survey by id
      description: Updates the corresponding survey
      operationId: updateSurveyById
      parameters:
        - name: id
          in: path
          description: ID of survey to update
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
    delete:
      tags:
        - survey
      summary: Delete survey by id
      description: Deletes the corresponding survey
      operationId: deleteSurveyById
      parameters:
        - name: id
          in: path
          description: ID of survey to update
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
  /survey/{id}/analysis:
    get:
      tags:
        - survey
      summary: Returns an analysis for the given survey
      description: Returns an analysis object that contains data about the survey participants
      operationId: getSurveyAnalysisById
      parameters:
        - name: id
          in: path
          description: ID of survey to return an analysis for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
  /survey/{id}/submissions:
    get:
      tags:
        - submission
      summary: Returns all submissions for the given survey
      description: Returns a list of all submissions for the given survey
      operationId: getSurveySubmissionsById
      parameters:
        - name: id
          in: path
          description: ID of survey to return all submissions for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Submission'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
  /survey/{id}/results:
    get:
      tags:
        - submission
      summary: Returns results of the given survey
      description: Returns a more simple representation of survey results for the end-user to view
      operationId: getSurveyResultsById
      parameters:
        - name: id
          in: path
          description: ID of survey to return results for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Results'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /survey/{id}/submission:
    post:
      tags:
        - submission
      summary: Creates a new submission for the given survey
      description: Endpoint is used by the end-user to create a new submission
      operationId: createSurveySubmission
      parameters:
        - name: id
          in: path
          description: ID of survey to create a new submission for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /participant/{id}/surveys:
    get:
      tags:
        - submission
      summary: Return all surveys for the logged in user
      description: All surveys for the corresponding user are returned
      operationId: getParticipantsSurveys
      parameters:
        - name: id
          in: path
          description: id of participant to fetch surveys from
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Survey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
components:
  schemas:
    User:
      required:
        - id
        - email
        - password
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
...
    Survey:
      required:
        - id
        - name
        - user
        - questions
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        user:
          $ref: '#/components/schemas/User'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    SurveyPrepare:
      required:
        - name
        - questions
      type: object
      properties:
        name:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    Question:
      required:
        - id
        - survey_id
        - text
      type: object
      properties:
        id:
          type: integer
          format: int64
        survey_id:
          type: integer
          format: int64
        text:
          type: string
        mode:
          type: string
          enum: [radio, check, text, rating, dropdown, nps, consent]
        answerOptions:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOption'
    AnswerOption:
      required:
        - id
        - question_id
        - content
      type: object
      properties:
        id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        content:
          oneOf:
            - type: string
            - type: number
            - type: integer
            - type: boolean
    Results:
      required:
        - mode
        - survey_id
      type: object
      properties:
        survey_id:
          type: integer
          format: int64
        mode:
          type: string
          enum: [radio, check, text, rating, dropdown, nps, consent]
        choices:
          type: array
          items:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
    Submission:
      required:
        - id
        - participant
        - survey_id
        - timestamp
        - answers
      type: object
      properties:
        id:
          type: integer
          format: int64
        participant:
          $ref: '#/components/schemas/Participant'
        survey_id:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Answer'
    Answer:
      required:
        - id
        - question_id
        - choices
      type: object
      properties:
        id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        choices:
          type: array
          items:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
    Participant:
      required:
        - id
        - ip_address
      type: object
      properties:
        id:
          type: integer
          format: int64
        ip_address:
          type: string
          format: ipv4
    Analysis:
      required:
        - id
        - survey_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        survey_id:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
          example: 527
        countries:
          type: object
          additionalProperties:
            type: integer
          example:
            Germany: 69
            USA: 420
            Spain: 11
            Norway: 27
...
    ApiError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: api_key
      in: header